A second version of the experiment process.
This is implemented to gain more controll over the process, so wrapper functions for the outer loop cv are not used.

Now we can visualize more information, for example:
	*  The best parameters for each learning method on each outer loop iteration.

Linear kernel implementation gets stuck at some point with the dataset D, so it was excluded from the hyperparameter grid of SVR.
LinearSVM should be used instead.